requirements.txt
# requirements.txt

pandas
numpy
scipy
pyyaml
pytest